Static PHP responsive website
Project Synopsis: You will create a completely responsive website, featuring a minimum of 4 pages, that uses a foundation-level php ‘include’ file template system. Additionally, you will be required to implement several custom PHP functions to automate portions of the site, as well as develop a basic content strategy for the topic you choose for the site design. Although this project is highly technical, you will also be graded significantly on visual design sophistication and relevance to the topic.